1、在直接系统调用的基础上升级为间接系统调用2、更新文档3、修复一个在Mingw上无法切换编译版本的Bug

+++ src/invoker/NtExtInvokers.hpp
@@ -16,13 +16,13 @@
 namespace NtExt {
     #ifdef _WIN64
     _Check_return_ inline X64Call Call(_In_ DWORD64 target) { return X64Call(target); }
-    _Check_return_ inline X64Syscall Syscall(_In_ WORD ssn) { return X64Syscall(ssn); }
+    _Check_return_ inline X64Syscall Syscall(_In_ DWORD64 ssn) { return X64Syscall(ssn); }
     _Check_return_ inline X64Anycall Anycall(_In_ const std::string& opcode) { return X64Anycall(opcode); }
     #endif
 
     #ifdef _M_IX86
     _Check_return_ inline Wow64Call Call(_In_ DWORD64 target) { return Wow64Call(target); }
-    _Check_return_ inline Wow64Syscall Syscall(_In_ WORD ssn) { return Wow64Syscall(ssn); }
+    _Check_return_ inline Wow64Syscall Syscall(_In_ DWORD64 ssn) { return Wow64Syscall(ssn); }
     _Check_return_ inline Wow64Anycall Anycall(_In_ const std::string& opcode) { return Wow64Anycall(opcode); }
     #endif
 }
